+++ activity_manager/src/main/java/exort/activity/entity/Select.java
@@ -3,7 +3,8 @@ package exort.activity.entity;
 import java.util.List;
 
 public class Select {
-    private Long associationId;
+
+    private int associationId;
     private String keyword;
     private List<String> tags;
     private TimeRange createTime;
@@ -15,11 +16,9 @@ public class Select {
     private int ifReview;
     private int ifOnlyMem;
 
-    public Long getAssociationId() {
-        return associationId;
-    }
+    public int getAssociationId() { return associationId; }
 
-    public void setAssociationId(Long associationId) {
+    public void setAssociationId(int associationId) {
         this.associationId = associationId;
     }
